+++ client/clientd/src/lib.rs
@@ -1,20 +1,32 @@
+use std::sync::Arc;
+
 use anyhow::Result;
 use async_trait::async_trait;
 use axum::extract::rejection::JsonRejection;
 use axum::extract::{FromRequest, RequestParts};
 use axum::response::{IntoResponse, Response};
+use axum::routing::post;
 use axum::BoxError;
+use axum::{Extension, Router, Server};
 use bitcoin::hashes::hex::ToHex;
+use bitcoin::secp256k1::rand;
 use bitcoin::Transaction;
 use fedimint_api::{Amount, OutPoint, TieredMulti, TransactionId};
 use fedimint_core::modules::wallet::txoproof::TxOutProof;
 use mint_client::mint::{NoteIssuanceRequests, SpendableNote};
 use mint_client::ClientError;
+use mint_client::{Client, UserClientConfig};
+use rand::rngs::OsRng;
 use reqwest::StatusCode;
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 use thiserror::Error;
+use tokio::sync::mpsc;
+use tokio::sync::mpsc::Sender;
+use tower::ServiceBuilder;
+use tower_http::trace::{DefaultMakeSpan, DefaultOnRequest, DefaultOnResponse, TraceLayer};
+use tracing::{info, Level};
 
 #[derive(Error, Debug)]
 pub enum ClientdError {
@@ -207,5 +219,136 @@ impl From<TieredMulti<SpendableNote>> for TieredNoteCount {
             })
             .collect();
         TieredNoteCount(all_tiered)
+    }
+}
+
+// TODO: generalize for all types of clients?
+struct State {
+    client: Arc<Client<UserClientConfig>>,
+    fetch_tx: Sender<()>,
+    rng: OsRng,
+}
+
+pub async fn run_clientd(client: Arc<Client<UserClientConfig>>, port: u16) -> Result<()> {
+    // make this a lib >>>>
+    let (tx, mut rx) = mpsc::channel(1024);
+    let rng = OsRng;
+
+    let shared_state = Arc::new(State {
+        client: Arc::clone(&client),
+        fetch_tx: tx,
+        rng,
+    });
+    let app = Router::new()
+        .route("/get_info", post(info))
+        .route("/get_pending", post(pending))
+        .route("/get_new_peg_in_address", post(new_peg_in_address))
+        .route("/wait_block_height", post(wait_block_height))
+        .route("/peg_in", post(peg_in))
+        .route("/spend", post(spend))
+        .layer(
+            ServiceBuilder::new()
+                .layer(
+                    TraceLayer::new_for_http()
+                        .make_span_with(DefaultMakeSpan::new().include_headers(true))
+                        .on_request(DefaultOnRequest::new().level(Level::INFO))
+                        .on_response(DefaultOnResponse::new().level(Level::INFO)),
+                )
+                .layer(Extension(shared_state)),
+        );
+
+    let fetch_client = Arc::clone(&client);
+    tokio::spawn(async move {
+        while rx.recv().await.is_some() {
+            fetch(Arc::clone(&fetch_client)).await;
+        }
+    });
+
+    Server::bind(&format!("127.0.0.1:{}", port).parse().unwrap())
+        .serve(app.into_make_service())
+        .await
+        .unwrap();
+    Ok(())
+}
+
+/// Handler for "get_info", returns all the clients holdings and pending transactions
+async fn info(Extension(state): Extension<Arc<State>>) -> Result<impl IntoResponse, ClientdError> {
+    let client = &state.client;
+    json_success!(InfoResponse::new(
+        client.coins(),
+        client.list_active_issuances(),
+    ))
+}
+
+/// Handler for "get_pending", returns the clients pending transactions
+async fn pending(
+    Extension(state): Extension<Arc<State>>,
+) -> Result<impl IntoResponse, ClientdError> {
+    let client = &state.client;
+    json_success!(PendingResponse::new(client.list_active_issuances()))
+}
+
+async fn new_peg_in_address(
+    Extension(state): Extension<Arc<State>>,
+) -> Result<impl IntoResponse, ClientdError> {
+    let client = &state.client;
+    let mut rng = state.rng;
+    json_success!(PegInAddressResponse {
+        peg_in_address: client.get_new_pegin_address(&mut rng)
+    })
+}
+
+async fn wait_block_height(
+    Extension(state): Extension<Arc<State>>,
+    Json(payload): Json<WaitBlockHeightPayload>,
+) -> Result<impl IntoResponse, ClientdError> {
+    let client = &state.client;
+    client.await_consensus_block_height(payload.height).await;
+    json_success!("done")
+}
+
+async fn peg_in(
+    Extension(state): Extension<Arc<State>>,
+    payload: Json<PegInPayload>,
+) -> Result<impl IntoResponse, ClientdError> {
+    let client = &state.client;
+    let fetch_signal = &state.fetch_tx;
+    let mut rng = state.rng;
+    let txout_proof = payload.0.txout_proof;
+    let transaction = payload.0.transaction;
+    let txid = client.peg_in(txout_proof, transaction, &mut rng).await?;
+    info!("Started peg-in {}", txid.to_hex());
+    fetch_signal
+        .send(())
+        .await
+        .map_err(|_| ClientdError::ServerError)?;
+    json_success!(PegInOutResponse { txid })
+}
+
+async fn spend(
+    Extension(state): Extension<Arc<State>>,
+    payload: Json<SpendPayload>,
+) -> Result<impl IntoResponse, ClientdError> {
+    let client = &state.client;
+    let rng = state.rng;
+
+    let notes = client.spend_ecash(payload.0.amount, rng).await?;
+    json_success!(SpendResponse { notes })
+}
+
+async fn fetch(client: Arc<Client<UserClientConfig>>) {
+    //TODO: log txid or error (handle unwrap)
+    let batch = client.fetch_all_coins().await;
+    for item in batch.iter() {
+        match item {
+            Ok(out_point) => {
+                //TODO: Log event
+                info!("fetched notes: {}", out_point);
+            }
+            Err(err) => {
+                //TODO: Log event
+                info!("error fetching notes: {}", err);
+            }
+        }
     }
 }
